Refactored endpoint for importing files
