+++ .gitignore
@@ -7,3 +7,4 @@ node_modules/
 .vscode/* 
 .DS_Store
 *.env
+backend/collaborative-content-coding-firebase-adminsdk-dpj86-0d9f3ad3d8.json
